% Generated by roxygen2 (4.0.2): do not edit by hand
\name{branch.and.bound.max.uncertainty}
\alias{branch.and.bound.max.uncertainty}
\title{Computes the maximum uncertainty by a branch and bound algorithm.}
\usage{
branch.and.bound.max.uncertainty(X, f.X, K.hat = NULL, tol = 0.001)
}
\arguments{
\item{X}{\eqn{X,}{} a matrix whose rows are points \eqn{\textrm{in } [0, 1]^p}{} where \eqn{f} is observed}

\item{f.X}{\eqn{f|_X,}{} a vector whose values are \eqn{f} evaluated at each row of \eqn{X}}

\item{K.hat}{\eqn{\hat{K},}{} the empirical Lipschitz constant of \eqn{f}}

\item{tol}{solutions this close to the maximum uncertainty are good enough. Defaults to 0.001.}
}
\description{
Computes the maximum uncertainty by a branch and bound algorithm.
}
\examples{
X <- expand.grid(1:9, 1:9) / 10
f <- function(x) sin(x[1]) + cos(x[2])
f.X <- apply(X, 1, f)
K.hat <- find.K.hat(X, f.X)
branch.and.bound.max.uncertainty(X, f.X, K.hat)
}

