% Generated by roxygen2 (4.0.2): do not edit by hand
\name{corners.uncertainty.bound}
\alias{corners.uncertainty.bound}
\title{Computes a lower bound on the maximum uncertainty by evaluating
the pointwise error at corners of the domain.}
\usage{
corners.uncertainty.bound(X, f.X, n = 5000, first.corner = 0,
  K.hat = NULL)
}
\arguments{
\item{X}{a finite subset of \eqn{[0, 1]^p} where \eqn{f} is observed}

\item{f.X}{\eqn{f|_X}, the values of \eqn{f} on \eqn{X}}

\item{n}{the number of consecutive corners to consider. Defaults to 5000.}

\item{first.corner}{the first corner to consider. Defaults to 0.}

\item{K.hat}{the empirical Lipschitz constant of \eqn{f}. Defaults to \code{find.K.hat(X, f.X)}.}
}
\description{
Computes a lower bound on the maximum uncertainty by evaluating
the pointwise error at corners of the domain.
}

