% Generated by roxygen2 (4.0.2): do not edit by hand
\name{find.K.hat}
\alias{find.K.hat}
\title{Finds the empirical Lipschitz constant of \eqn{f},
a lower bound on the Lipschitz constant of \eqn{f}.}
\usage{
find.K.hat(X, f.X, lp.norm = Inf, min.distance = 0)
}
\arguments{
\item{X}{a finite subset of \eqn{[0, 1]^p} where \eqn{f} is observed}

\item{f.X}{\eqn{f|_X}, the values of \eqn{f} on \eqn{X}}

\item{lp.norm}{the \eqn{L_p} norm for computing the distances between points in \eqn{X}. Defaults to sup-norm.}

\item{min.distance}{only pairs of points in \eqn{X} farther apart than this threshold are considered. Defaults to 0.}
}
\description{
Finds the empirical Lipschitz constant of \eqn{f},
a lower bound on the Lipschitz constant of \eqn{f}.
}

