% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lower.bound.max.uncertainty}
\alias{lower.bound.max.uncertainty}
\title{Computes a lower bound on the maximum uncertainty within the specified box,
by evaluating the pointwise error at the box's corners and at random points
in the box's interior.}
\usage{
lower.bound.max.uncertainty(X, f.X, K.hat, box = NULL)
}
\arguments{
\item{X}{a finite subset of \eqn{[0, 1]^p} where \eqn{f} is observed}

\item{f.X}{\eqn{f|_X}, the values of \eqn{f} on \eqn{X}}

\item{K.hat}{the empirical Lipschitz constant of \eqn{f}}

\item{box}{the top-left corner and the bottom-right corner of a hypercube contained in the domain. Defaults to \eqn{[0,1]^p}.}
}
\description{
Computes a lower bound on the maximum uncertainty within the specified box,
by evaluating the pointwise error at the box's corners and at random points
in the box's interior.
}

