% Generated by roxygen2 (4.0.2): do not edit by hand
\name{pointwise.uncertainty}
\alias{pointwise.uncertainty}
\title{Computes pointwise uncertainty}
\usage{
pointwise.uncertainty(X, f.X, K.hat, x)
}
\arguments{
\item{X}{a finite subset of \eqn{[0, 1]^p} where \eqn{f} is observed}

\item{f.X}{\eqn{f|_X}, the values of \eqn{f} on \eqn{X}}

\item{K.hat}{the empirical Lipschitz constant of \eqn{f}}

\item{x}{the point for which to compute pointwise uncertainty}
}
\description{
Computes pointwise uncertainty
}

