% Generated by roxygen2 (4.0.2): do not edit by hand
\name{branch.and.bound.max.uncertainty}
\alias{branch.and.bound.max.uncertainty}
\title{Computes the maximum uncertainty by a branch and bound algorithm.}
\usage{
branch.and.bound.max.uncertainty(X, f.X, K.hat = NULL, tol = 0.001)
}
\arguments{
\item{X}{a finite subset of \eqn{[0, 1]^p} where \eqn{f} is observed}

\item{f.X}{\eqn{f|_X}, the values of \eqn{f} on \eqn{X}}

\item{K.hat}{the empirical Lipschitz constant of \eqn{f}}

\item{tol}{solutions this close to the maximum uncertainty are good enough. Defaults to 0.001.}
}
\description{
Computes the maximum uncertainty by a branch and bound algorithm.
}

