% Generated by roxygen2 (4.0.2): do not edit by hand
\name{corners.uncertainty.bound}
\alias{corners.uncertainty.bound}
\title{Computes a lower bound on the maximum uncertainty by evaluating
the pointwise error at corners of the domain.}
\usage{
corners.uncertainty.bound(X, f.X, K.hat, n = 5000, first.corner = 0)
}
\arguments{
\item{X}{\eqn{X,}{} a matrix whose rows are points \eqn{\textrm{in } [0, 1]^p}{} where \eqn{f} is observed}

\item{f.X}{\eqn{f|_X,}{} a vector whose values are \eqn{f} evaluated at each row of \eqn{X}}

\item{K.hat}{\eqn{\hat{K},}{} the empirical Lipschitz constant of \eqn{f}}

\item{n}{the number of consecutive corners to consider. Defaults to 5000.}

\item{first.corner}{the first corner to consider. Defaults to 0.}
}
\description{
Computes a lower bound on the maximum uncertainty by evaluating
the pointwise error at corners of the domain.
}
\examples{
X <- expand.grid(1:9, 1:9) / 10
f <- function(x) sin(x[1]) + cos(x[2])
f.X <- apply(X, 1, f)
K.hat <- find.K.hat(X, f.X)
corners.uncertainty.bound(X, f.X, K.hat)
}

