% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lower.bound.computational.burden}
\alias{lower.bound.computational.burden}
\title{Computes a lower bound on the minimum computational burden.}
\usage{
lower.bound.computational.burden(X, f.X, epsilon, lp.norm = Inf,
  K.hat = NULL)
}
\arguments{
\item{X}{a finite subset of \eqn{[0, 1]^p} where \eqn{f} is observed}

\item{f.X}{\eqn{f|_X}, the values of \eqn{f} on \eqn{X}}

\item{epsilon}{the maximum uncertainty allowed}

\item{lp.norm}{the \eqn{L_p} norm for computing the distances between points in \eqn{[0, 1]^p}. Defaults to sup-norm.}

\item{K.hat}{the empirical Lipschitz constant of \eqn{f}. Defaults to find.K.hat(X, f.X).}
}
\description{
Computes a lower bound on the minimum computational burden.
}

