% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lower.bound.max.uncertainty}
\alias{lower.bound.max.uncertainty}
\title{Computes a lower bound on the maximum uncertainty within the specified box,
by evaluating the pointwise error at the box's corners and at random points
in the box's interior.}
\usage{
lower.bound.max.uncertainty(X, f.X, K.hat, box = NULL)
}
\arguments{
\item{X}{\eqn{X,}{} a matrix whose rows are points \eqn{\textrm{in } [0, 1]^p}{} where \eqn{f} is observed}

\item{f.X}{\eqn{f|_X,}{} a vector whose values are \eqn{f} evaluated at each row of \eqn{X}}

\item{K.hat}{\eqn{\hat{K},}{} the empirical Lipschitz constant of \eqn{f}}

\item{box}{the top-left corner and the bottom-right corner of a hypercube contained in the domain. Defaults to \eqn{[0,1]^p}.}
}
\description{
Computes a lower bound on the maximum uncertainty within the specified box,
by evaluating the pointwise error at the box's corners and at random points
in the box's interior.
}
\examples{
X <- expand.grid(1:9, 1:9) / 10
f <- function(x) sin(x[1]) + cos(x[2])
f.X <- apply(X, 1, f)
K.hat <- find.K.hat(X, f.X)
lower.bound.max.uncertainty(X, f.X, K.hat)
}

