\name{MiniMiniMaxUQ-package}
\alias{MiniMiniMaxUQ-package}
\alias{MiniMiniMaxUQ}
\docType{package}
\title{
Mini-Minimax Uncertainty Quantification for Emulators
}
\description{
MiniMiniMaxUQ seeks to answer the following: How
    accurately can a black-box function f be emulated from a given set of data?
    How many evaluations of f are required to guarantee that f can be emulated
    to a given level of accuracy?
}
\details{
\tabular{ll}{
Package: \tab MiniMiniMaxUQ\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-20\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Jeffrey C. Regier and Philip B. Stark

Maintainer: Jeffrey Regier <jeff@stat.berkeley.edu>
}
\references{
Jeffrey C. Regier and Philip B. Stark. "Mini-Minimax Uncertainty Quantification for Emulators." arXiv preprint arXiv:1303.3079 (2013).
}
\keyword{ package }
\seealso{
}
\examples{
}
