% Generated by roxygen2 (4.0.2): do not edit by hand
\name{uncertainty.confidence.bounds}
\alias{uncertainty.confidence.bounds}
\title{Computes confidence bounds for 4 statistics (first quartile, median, third quartile, mean)
of the distribution of pointwise uncertainties.}
\usage{
uncertainty.confidence.bounds(X, f.X, K.hat, n = 1000, confidence = 0.95)
}
\arguments{
\item{X}{a finite subset of \eqn{[0, 1]^p} where \eqn{f} is observed}

\item{f.X}{\eqn{f|_X}, the values of \eqn{f} on \eqn{X}}

\item{K.hat}{the empirical Lipschitz constant of \eqn{f}}

\item{n}{the number of samples. Defaults to 1000.}

\item{confidence}{the size of the confidence bounds. Defaults to 0.95.}
}
\description{
Computes confidence bounds for 4 statistics (first quartile, median, third quartile, mean)
of the distribution of pointwise uncertainties.
}

